Split UISheet and rename constants
