+++ packages/navigation/src/Sheets/UISheet/types.ts
@@ -1,0 +1,1 @@
+export type OnClose = () => void | Promise<void>;
