+++ packages/navigation/src/Sheets/index.tsx
@@ -1,2 +1,4 @@
-export * from './UISheet';
+export * from './UISheet/UISheet';
 export * from './UIQRCodeScannerSheet';
+export * from './UICardSheet';
+export * from './UIBottomSheet';
