+++ packages/navigation/src/UISlideBar.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, View } from 'react-native';
 import { UIConstant } from '@tonlabs/uikit.core';
 import { ColorVariants, UIBackgroundView } from '@tonlabs/uikit.hydrogen';
 
-import { HEADER_HEIGHT } from './constants';
+import { UIConstant as UINavConstant } from './constants';
 import { UIHeaderItems } from './UIHeaderItems';
 import type { HeaderItem } from './UIHeaderItems';
 import { useNavigationHeaderLeftItems } from './useNavigationHeaderLeftItems';
@@ -71,7 +71,7 @@ export function UISlideBar({
 
 const styles = StyleSheet.create({
     container: {
-        minHeight: HEADER_HEIGHT,
+        minHeight: UINavConstant.headerHeight,
         flexDirection: 'row',
         alignItems: 'center',
         paddingVertical: UIConstant.smallContentOffset(),
