+++ packages/navigation/src/constants.tsx
@@ -1,25 +1,25 @@
-export const HEADER_HEIGHT = 56;
-export const ELASTIC_WIDTH_CONTROLLER = 600;
-
-export const ICON_SEARCH_SIZE = 20;
-export const ICON_SEARCHING_INDICATOR_SIZE = 14;
-
-export const SCREEN_CONTENT_INSET_HORIZONTAL = 16;
 const GRID_LINE_HEIGHT = 4;
-export const CONTENT_INSET_VERTICAL_X2 = 2 * GRID_LINE_HEIGHT;
-export const CONTENT_INSET_VERTICAL_X4 = 4 * GRID_LINE_HEIGHT;
-
-export const CONTENT_OFFSET: number = 16;
-export const PAGER_VIEW_HEIGHT: number = 72;
-
-export const TITLE_MINIMUM_FONT_SCALE: number = 0.7;
 
 export const UIConstant = {
+    iconSearchSize: 20,
+    iconSearchingIndicatorSize: 14,
+
+    scrollContentInsetHorizontal: 16,
+    contentInsetVerticalX2: 2 * GRID_LINE_HEIGHT,
+    contentInsetVerticalX4: 4 * GRID_LINE_HEIGHT,
+
     contentOffset: 16,
     alertBorderRadius: 12,
 
+    headerHeight: 56,
+    pagerViewHeight: 72,
+
+    elastiicWidthController: 600,
+
     elasticWidthCardSheet: 414,
     elasticWidthBottomSheet: 448,
-
     swipeThreshold: 50,
+    rubberBandEffectDistance: 50,
+
+    titleMinimumFontScale: 0.7,
 };
